Deleted some files, use pragma once instead of ifndef...

+++ src/client/client.cpp
@@ -182,7 +182,7 @@ int cumulativeMouseMovement = 0;
 int clientPortStart = 0; /* First UDP port for clients */
 int clientPortEnd = 0;   /* Last one (these are for firewalls) */
 
-byte lose_item;       /* index for dropping owned item */
+int lose_item;        /* index for dropping owned item */
 int lose_item_active; /* one of the lose keys is pressed */
 
 #ifdef SOUND
